Backend implemented, DB filled
Implemented a Node.js/Express.js backend, with connection to MongoDB.
Created a fantasy dataset.

Connected the frontend and tweaked the flow of data a little.
The shopping cart has a seperate call, to ensure the latest data on checkout.

ToDo:

My approach of only having an API call on the Homepage, doesnt add up in the end.
If something happens to the Redux-State, the whole site breaks.

Ideas:
- Implementing a check, if the redux state is filled, if not, redirect to home to get the latest state
- Have extra API calls on categoryPage and itemPage, to get the latest data. But this should not really word, because without a redux state, i will not have am id to check... we'll see.