Found an ok solution for image loading
I tried several techniques to achieve lazy loading for the images.
The 'problem' is, that the images change size dynamically with the responsive layout.

Lazy loading only really works with fixed height and width, which i dont have.

I tried a package 'react-lazy-loading-image' but the result was the same.

So i stick with my custom solution. Now the user gets indicated, that there will be an image soon, but the height does not reflect the actual height of the image 100% of the time.